[#16] refactor: 파일명 명시적으로 변경
